--- DataRenkei/DataRenkei/Models/viewModel.swift
@@ -1,14 +1,0 @@
-//
-//  viewModel.swift
-//  DataRenkei
-//
-//  Created by 健太郎 on 2024/05/19.
-//
-
-import Foundation
-
-struct UserName: Identifiable {
-    var id = UUID()  // 一意の識別子
-    var name: String
-}
-
