+++ data_renkei_demo/data_renkei_demo.xcodeproj/xcuserdata/kentaro.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,7 +4,7 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>DataRenkei.xcscheme_^#shared#^_</key>
+		<key>data_renkei_demo.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>0</integer>
