+++ data_renkei_demo/data_renkei_demo/Assets.xcassets/AccentColor.colorset/Contents.json
