+++ data_renkei_demo/data_renkei_demo/Assets.xcassets/AppIcon.appiconset/Contents.json
