+++ data_renkei_demo/data_renkei_demo/ContentView.swift
@@ -1,0 +1,24 @@
+//
+//  ContentView.swift
+//  data_renkei_demo
+//
+//  Created by 健太郎 on 2024/05/26.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        VStack {
+            Image(systemName: "globe")
+                .imageScale(.large)
+                .foregroundStyle(.tint)
+            Text("Hello, world!")
+        }
+        .padding()
+    }
+}
+
+#Preview {
+    ContentView()
+}
