+++ data_renkei_demo/data_renkei_demo/data_renkei_demoApp.swift
@@ -1,0 +1,17 @@
+//
+//  data_renkei_demoApp.swift
+//  data_renkei_demo
+//
+//  Created by 健太郎 on 2024/05/26.
+//
+
+import SwiftUI
+
+@main
+struct data_renkei_demoApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
